[bndtools] ConcurrentModificationException occurs during: "Create implicit repo"
fixes #3819
