+++ bndtools.m2e/src/bndtools/m2e/MavenWorkspaceRunListener.java
@@ -1,7 +1,6 @@
 package bndtools.m2e;
 
 import org.bndtools.api.RunListener;
-import org.bndtools.api.RunMode;
 import org.eclipse.core.resources.IResource;
 import org.osgi.service.component.annotations.Component;
 
@@ -13,26 +12,6 @@ public class MavenWorkspaceRunListener implements MavenRunListenerHelper, RunLis
 
 	@Override
 	public void create(Run run) throws Exception {
-		IResource resource = getResource(run);
-
-		if (!isMavenProject(resource)) {
-			return;
-		}
-
-		Workspace workspace = run.getWorkspace();
-
-		RunMode runMode = RunMode.get(run);
-
-		if ((runMode == RunMode.LAUNCH) || (runMode == RunMode.TEST)) {
-			MavenWorkspaceRepository repo = workspace.getPlugin(MavenWorkspaceRepository.class);
-
-			if (repo == null) {
-				repo = new MavenWorkspaceRepository();
-				workspace.getRepositories()
-					.add(0, repo);
-				workspace.addBasicPlugin(repo);
-			}
-		}
 	}
 
 	@Override
@@ -45,17 +24,10 @@ public class MavenWorkspaceRunListener implements MavenRunListenerHelper, RunLis
 
 		Workspace workspace = run.getWorkspace();
 
-		RunMode runMode = RunMode.get(run);
+		MavenWorkspaceRepository repo = workspace.getPlugin(MavenWorkspaceRepository.class);
 
-		if ((runMode == RunMode.LAUNCH) || (runMode == RunMode.TEST)) {
-			MavenWorkspaceRepository repo = workspace.getPlugin(MavenWorkspaceRepository.class);
-
-			if (repo != null) {
-				workspace.getRepositories()
-					.remove(repo);
-				workspace.removeBasicPlugin(repo);
-				repo.cleanup();
-			}
+		if (repo != null) {
+			repo.cleanup();
 		}
 	}
 }
